Avoid printing an error when rerunning build-compiler-rt.sh on msys2

+++ build-compiler-rt.sh
@@ -39,7 +39,9 @@ fi
 # Add a symlink for i386 -> i686; we normally name the toolchain
 # i686-w64-mingw32, but due to the compiler-rt cmake peculiarities, we
 # need to refer to it as i386 at this stage.
-ln -sfn i686-w64-mingw32 $PREFIX/i386-w64-mingw32 || true
+if [ ! -e $PREFIX/i386-w64-mingw32 ]; then
+    ln -sfn i686-w64-mingw32 $PREFIX/i386-w64-mingw32 || true
+fi
 
 cd compiler-rt
